fix: container delete

+++ src/view/mate/view/Activities.ts
@@ -143,6 +143,8 @@ export default class Activities implements View {
     }
 
     public close(): void {
-        this.container.delete();
+        if (this.container.deleted !== true) {
+            this.container.delete();
+        }
     }
 }
