+++ src/view/mate/view/Faq.ts
@@ -28,6 +28,8 @@ export default class Faq implements View {
     }
 
     public close(): void {
-        this.container.delete();
+        if (this.container.deleted !== true) {
+            this.container.delete();
+        }
     }
 }
