+++ src/view/mate/view/mates/FollowMe.ts
@@ -36,6 +36,8 @@ export default class FollowMe implements View {
     }
 
     public close(): void {
-        this.container.delete();
+        if (this.container.deleted !== true) {
+            this.container.delete();
+        }
     }
 }
